Add files via upload
This is a PDF version of the Maple file corresponding to Sections 2.2.2 and 4.3 of the project "Ulam_sequences.pdf". It includes further work not exposed in the project. It was prepared for a presentation and meant to be accompanied by an explanation. It is not meant to be understood on its own and will be more comprehensible after reading "Ulam_sequences.pdf " and alongside "Equidistribution of Ulam sequences (presentation).ipynb".